pyproject.toml
[project]
name = "compiladores"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "antlr4-tools>=0.2.2",
    "transcrypt>=3.9.4",
]
